Компановка Email функционала
