+++ email/18-ISbo-2b/main_5_InformUsers.py
@@ -1,11 +1,10 @@
 # coding=utf-8
 from work_Loger import Logs
 from email.message import EmailMessage
+import work_EmailLibrary as EmailLibrary
 import config_Project as cfg
 import config_Mail
 
-import smtplib
-import email
 from datetime import datetime
 import inspect
 
@@ -16,14 +15,14 @@ def InformUsers(answersForUsers):
     """
 
     # Создание SMTP объекта
-    smtp_obj = smtp_login()
+    smtp_obj = EmailLibrary.smtp_login()
 
     # Отправка ответов пользователям
     # Глобальная функция 9
     SendLetters(smtp_obj, answersForUsers)
 
     # Закрытие SMTP объекта
-    quit_email_smtp(smtp_obj)
+    EmailLibrary.quit_email_smtp(smtp_obj)
 
     # Формирование нового имени файла логов
     FormFilename()
@@ -49,34 +48,7 @@ def SendLetters(smtp_obj, answersForUsers):
 
     for i in answersForUsers:
         # Отправка ответа по экземпляру списка ответов
-        send_mes(smtp_obj, i)
-
-
-def send_mes(smtp_obj, message):
-    try:
-        # Создание экземпляра класса Email Message
-        mes = EmailMessage()
-
-        # Заполнение поля отправителя
-        mes['From'] = "ТРПО ИАСТ"
-
-        # Заполнение поля получателя
-        mes['To'] = message.Who
-
-        # Заполнение темы письма
-        mes['Subject'] = message.Theme
-
-        # Заполнение тела письма
-        mes.set_content(message.Body)
-
-        # отправка SMTP пакета
-        smtp_obj.send_message(mes)
-        
-        message.Success = True
-
-    except:
-        message.Success = False
-
+        EmailLibrary.send_mes(smtp_obj, i)
 
 def FormFilename():
     """
@@ -90,28 +62,3 @@ def FormFilename():
         cfg.gen_num_for_filename = cfg.num_for_filename()
 
     cfg.filename = cfg.path_to_logs + "log_" + name + "_" + str(next(cfg.gen_num_for_filename)) + ".txt"
-
-def smtp_login():
-    """
-    Авторизация в Gmail аккаунте.
-    Функция возвращает SMTP объект.
-    :return:
-    """
-
-    smtpObj = smtplib.SMTP('smtp.gmail.com:587')
-    smtpObj.ehlo()
-    smtpObj.starttls()
-    smtpObj.ehlo()
-    smtpObj.login(config_Mail.EMAIL_ADDRESS, config_Mail.EMAIL_PASSWORD)
-
-    return smtpObj
-
-def quit_email_smtp(smtpObj):
-    """
-    Закрытие SMTP объекта.
-    Функция должна быть вызвана после завершения рыботы с SMTP объектом.
-    :param smtpObj:
-    :return:
-    """
-
-    smtpObj.close()
